# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Documentation",
      "version": "1.0.0"
    },
    "paths": {
      "/movies/search": {
        "get": {
          "summary": "Search movies",
          "tags": ["Movies"],
          "operationId": "getAllMovies",
          "parameters": [
            {
              "name": "title",
              "in": "query",
              "description": "Movie title",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "year",
              "in": "query",
              "description": "Movie year",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "page",
              "in": "query",
              "description": "Page number",
              "required": false,
              "schema": {
                "type": "integer",
                "minimum": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Movies retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "title": {
                              "type": "string"
                            },
                            "year": {
                              "type": "string"
                            },
                            "imdbID": {
                              "type": "string"
                            },
                            "imdbRating": {
                              "type": "number"
                            },
                            "rottenTomatoesRating": {
                              "type": "integer"
                            },
                            "metacriticRating": {
                              "type": "integer"
                            },
                            "classification": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "pagination": {
                        "type": "object",
                        "properties": {
                          "total": {
                            "type": "integer"
                          },
                          "lastPage": {
                            "type": "integer"
                          },
                          "perPage": {
                            "type": "integer"
                          },
                          "currentPage": {
                            "type": "integer"
                          },
                          "from": {
                            "type": "integer"
                          },
                          "to": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameter",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Invalid query parameter"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },      
      "/movie/data/{imdbID}": {
        "get": {
          "summary": "Get movie data by ID",
          "tags": ["Movies"],
          "operationId": "getMovieById",
          "parameters": [
            {
              "name": "imdbID",
              "in": "path",
              "description": "IMDb ID of the movie",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "year": {
                        "type": "string"
                      },
                      "runtime": {
                        "type": "string"
                      },
                      "genres": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "country": {
                        "type": "string"
                      },
                      "principals": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "category": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "characters": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Invalid query parameter"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Movie not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default": "Movie not found"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "default": true
                      },
                      "message": {
                        "type": "string",
                        "default":"Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }